requirements.txt
asgiref==3.7.2
attrs==23.1.0
build==0.10.0
CacheControl==0.12.14
certifi==2023.5.7
charset-normalizer==3.2.0
cleo==2.0.1
colorama==0.4.6
crashtest==0.4.1
crispy-bootstrap4==2022.1
distlib==0.3.7
dj-database-url==2.0.0
Django==4.2.2
django-crispy-forms==2.0
django-environ==0.10.0
dulwich==0.21.5
filelock==3.12.2
gunicorn==20.1.0
html5lib==1.1
idna==3.4
importlib-metadata==6.8.0
installer==0.7.0
jaraco.classes==3.3.0
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
keyring==23.13.1
lockfile==0.12.2
more-itertools==9.1.0
msgpack==1.0.5
packaging==23.1
pdfkit==1.0.0
pexpect==4.8.0
Pillow==9.5.0
pip==23.2.1
pkginfo==1.9.6
platformdirs==3.9.1
poetry==1.5.1
poetry-core==1.6.1
poetry-plugin-export==1.4.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pypng==0.20220715.0
pyproject_hooks==1.0.0
pywin32-ctypes==0.2.2
qrcode==7.4.2
rapidfuzz==2.15.1
referencing==0.30.0
reportlab==4.0.4
requests==2.31.0
requests-toolbelt==1.0.0
rpds-py==0.9.2
setuptools==67.8.0
shellingham==1.5.0.post1
six==1.16.0
sqlparse==0.4.4
tomlkit==0.11.8
trove-classifiers==2023.7.6
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
virtualenv==20.24.1
webencodings==0.5.1
wheel==0.40.0
whitenoise==6.5.0
xlwt==1.3.0
zipp==3.16.2
django-simple-history==3.3.0
python-decouple==3.8
django-import-export==3.2.0